feat: move themecontext to core
