+++ apps/core/src/contexts/index.ts
@@ -1,4 +1,4 @@
 // Copyright (c) 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './PopupContext';
+export * from './ThemeContext';
