+++ apps/wallet-dashboard/app/(protected)/layout.tsx
@@ -9,11 +9,16 @@ import { Button } from '@iota/apps-ui-kit';
 import { redirect } from 'next/navigation';
 import { Sidebar } from './components';
 import { TopNav } from './components/top-nav/TopNav';
-import { useTheme } from '@/contexts';
+import { Theme, useTheme } from '@iota/core';
 
 function DashboardLayout({ children }: PropsWithChildren): JSX.Element {
     const { connectionStatus } = useCurrentWallet();
-    const { theme, toggleTheme } = useTheme();
+    const { theme, setTheme } = useTheme();
+
+    const toggleTheme = () => {
+        const newTheme = theme === Theme.Light ? Theme.Dark : Theme.Light;
+        setTheme(newTheme);
+    };
     const account = useCurrentAccount();
     useEffect(() => {
         if (connectionStatus !== 'connected' && !account) {
