+++ apps/wallet-dashboard/components/staking-overview/StartStaking.tsx
@@ -2,9 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { Button, ButtonSize, ButtonType, Panel } from '@iota/apps-ui-kit';
-import { Theme, useTheme } from '@/contexts';
 import { useState } from 'react';
 import { StakeDialog } from '../Dialogs';
+import { Theme, useTheme } from '@iota/core';
 
 export function StartStaking() {
     const { theme } = useTheme();
